Updated the upgrade system to be smoother and easier to extend.
However I still don't fully understand the relationship between tile entities and blocks. I could find no explicit creation of a tile entity when a block is placed. What creates that object? What holds the reference to it. Why doesn't the block itself hold the reference to it's tile entity?